# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Taco gaming",
        "description": "Apis",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://code-connoisseurs-laravel.vercel.app/rest"
        }
    ],
    "paths": {
...
        "/usuarios/{id}/compras": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Devuelve las compras realizadas por un usuario",
                "operationId": "cea27edd79bb41966e66e9d21d3d9146",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compras del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID de la compra",
                                                "type": "integer"
                                            },
                                            "fecha_compra": {
                                                "description": "Fecha de la compra",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "juegos_comprados": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "description": "ID del juego",
                                                            "type": "integer"
                                                        },
                                                        "titulo": {
                                                            "description": "Título del juego",
                                                            "type": "string"
                                                        },
                                                        "descripcion": {
                                                            "description": "Descripción del juego",
                                                            "type": "string"
                                                        },
                                                        "precio": {
                                                            "description": "Precio del juego",
                                                            "type": "number",
                                                            "format": "float"
                                                        },
                                                        "desarrollador": {
                                                            "properties": {
                                                                "id": {
                                                                    "description": "ID del desarrollador",
                                                                    "type": "integer"
                                                                },
                                                                "nombre": {
                                                                    "description": "Nombre del desarrollador",
                                                                    "type": "string"
                                                                },
                                                                "descripcion": {
                                                                    "description": "Descripción del desarrollador",
                                                                    "type": "string"
                                                                },
                                                                "pais": {
                                                                    "description": "País del desarrollador",
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuarios": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Registra un nuevo usuario.",
                "operationId": "0c522654a256aec9dbabac920b0a6393",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "Juan"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado de forma exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "usuario registrado de forma exitosa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "El nombre no puede ser vacío."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}